+++ Node/DynamicNode.cs
@@ -27,7 +27,7 @@ using UnityEngine;
 
 namespace Exodrifter.NodeGraph
 {
-	[Node(Name = "Dynamic Node", Path = "Dynamic")]
+	[Node(Name = "Dynamic Node", Path = "Other/Dynamic")]
 	public class DynamicNode : Node
 	{
 		[SerializeField]