+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * Added the ability to drag and drop objects into the node graph
+* Added the `NodeGraphPolicy` class, which defines what operations a node
+  graph can take
 
 ### Changed
 * Fuzzy search is now exhaustive, so it can return better results